rename store/actions.ts to write.ts
