+++ packages/webapp/src/components/modals/joinGameModal.tsx
@@ -10,7 +10,7 @@ import { isStringPositiveInteger, parseBigIntOrNull } from "src/utils/js-utils"
 import { Modal, ModalController, useModalController } from "src/components/lib/modal"
 import { LoadingModalContent } from "src/components/lib/loadingModal"
 import { joinGame, reportInconsistentGameState } from "src/actions"
-import { setError } from "src/store/actions"
+import { setError } from "src/store/write"
 import { GameStatus } from "src/store/types"
 import { navigate } from "src/utils/navigate"
 import { useCancellationHandler } from "src/hooks/useCancellationHandler"
