+++ packages/webapp/src/pages/play.tsx
@@ -14,7 +14,7 @@ import { readContract } from "wagmi/actions"
 import { deployment } from "src/deployment"
 import { gameABI } from "src/generated"
 import { useRouter } from "next/router"
-import { setError } from "src/store/actions"
+import { setError } from "src/store/write"
 import { DISMISS_BUTTON } from "src/actions/errors"
 import { navigate } from "src/utils/navigate"
 import { drawCard } from "src/actions/drawCard"
