+++ packages/webapp/src/store/update.ts
@@ -16,7 +16,7 @@ import * as net from "src/store/network"
 import { THROTTLED, ZOMBIE } from "src/utils/throttledFetch"
 import { formatTimestamp, parseBigInt } from "src/utils/js-utils"
 import { GameStatus } from "src/store/types"
-import { setError } from "src/store/actions"
+import { setError } from "src/store/write"
 import { contractWriteThrowing } from "src/actions/libContractWrite"
 import { deployment } from "src/deployment"
 import { gameABI } from "src/generated"
